backend/ml_service/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
numpy
git+https://github.com/ultralytics/ultralytics
